+++ R/expr-datetime.R
@@ -725,7 +725,8 @@ expr_dt_timestamp <- function(time_unit = c("us", "ns", "ms")) {
 #' Set time unit of a Series of dtype Datetime or Duration
 #' @description
 #' `r lifecycle::badge("deprecated")`
-#' Cast to Int64 and then to Datetime instead.
+#' Cast to Int64 and then to the desired Datetime or Duration dtype and time
+#' unit instead.
 #'
 #' @inheritParams expr_dt_timestamp
 #' @inherit as_polars_expr return
@@ -735,7 +736,10 @@ expr_dt_with_time_unit <- function(time_unit = c("ns", "us", "ms")) {
     deprecate_warn(
       c(
         `!` = "`$dt$with_time_unit()` is deprecated.",
-        i = "Cast to Int64 and then to Datetime with the desired time unit instead."
+        i = paste0(
+          "Cast to Int64 and then to the desired Datetime or Duration dtype ",
+          "and time unit instead."
+        )
       )
     )
     time_unit <- arg_match0(time_unit, values = c("ns", "us", "ms"))